backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.10.0
qdrant-client==1.16.2
google-generativeai==0.4.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.21.1